Complete Floor-70 and checkpoint main.dol progress

+++ include/effect/n64/eff_peach_dust_n64.h
@@ -2,4 +2,4 @@
 
 #include <dolphin/types.h>
 
-void* effPeachDustN64Entry(f32 x, f32 y, f32 z, s32 unused, f32 angle, s32 type);
+void* effPeachDustN64Entry(f32 x, f32 y, f32 z, f32 unused, f32 angle, s32 type);
